+++ src/TwigComponent/tests/Unit/CVATest.php
@@ -29,6 +29,29 @@ class CVATest extends TestCase
         $this->assertEquals($expected, $recipeClass->resolve($recipes));
     }
 
+    public function testApply(): void
+    {
+        $recipe = new CVA('font-semibold border rounded', [
+            'colors' => [
+                'primary' => 'text-primary',
+                'secondary' => 'text-secondary',
+            ],
+            'sizes' => [
+                'sm' => 'text-sm',
+                'md' => 'text-md',
+                'lg' => 'text-lg',
+            ],
+        ], [
+            [
+                'colors' => ['primary'],
+                'sizes' => ['sm'],
+                'class' => 'text-red-500',
+            ],
+        ]);
+
+        $this->assertEquals('font-semibold border rounded text-primary text-sm text-red-500', $recipe->apply(['colors' => 'primary', 'sizes' => 'sm']));
+    }
+
     public static function recipeProvider(): iterable
     {
         yield 'base null' => [